requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.9.0=py36_0
appnope=0.1.0=py36_1000
astor=0.7.1=py_0
attrs=19.3.0=py_0
backcall=0.1.0=py_0
bleach=3.1.1=py_0
c-ares=1.15.0=h01d97ff_1001
ca-certificates=2019.11.28=hecc5488_0
cctools=927.0.2=h5ba7a2e_4
certifi=2019.11.28=py36_0
clang=9.0.1=default_hf57f61e_0
clang_osx-64=9.0.1=h05bbb7f_0
clangxx=9.0.1=default_hf57f61e_0
clangxx_osx-64=9.0.1=h05bbb7f_0
cloudpickle=1.3.0=py_0
compiler-rt=9.0.1=h6a512c6_3
compiler-rt_osx-64=9.0.1=h99342c6_3
cycler=0.10.0=py_2
cytoolz=0.10.1=py36h0b31af3_0
dask-core=2.11.0=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py36_1000
freetype=2.10.0=h24853df_1
gast=0.3.3=py_0
grpcio=1.23.0=py36h8a88325_0
h5py=2.9.0=py36hafca783_1000
hdf5=1.10.4=nompi_h0cbb7df_1106
icu=58.2=h0a44026_1000
imageio=2.8.0=py_0
importlib_metadata=1.5.0=py36_0
ipykernel=5.1.4=py36h5ca1d4c_0
ipython=7.13.0=py36h5ca1d4c_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.5.1=py_0
jedi=0.16.0=py36_0
jinja2=2.11.1=py_0
jpeg=9c=h1de35cc_1001
jsonschema=3.2.0=py36_0
jupyter=1.0.0=py_2
jupyter_client=6.0.0=py_0
jupyter_console=5.1.0=py36_0
jupyter_core=4.6.3=py36_0
keras=2.3.1=py36_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_0
kiwisolver=1.1.0=py36ha1b3eb9_0
ld64=450.3=h3c32e8a_4
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libcxx=9.0.1=1
libffi=3.2.1=h6de7cb9_1006
libgfortran=3.0.1=0
libgpuarray=0.7.6=h1de35cc_1003
liblapack=3.8.0=14_openblas
libllvm9=9.0.1=ha1b3eb9_0
libopenblas=0.3.7=hd44dcd8_1
libpng=1.6.37=h2573ce8_0
libprotobuf=3.11.4=hd174df1_0
libsodium=1.0.17=h01d97ff_0
libtiff=4.1.0=ha78913b_3
llvm-openmp=8.0.1=h770b8ee_0
lz4-c=1.8.3=h6de7cb9_1001
mako=1.1.0=py_0
markdown=3.2.1=py_0
markupsafe=1.1.1=py36h0b31af3_0
matplotlib-base=3.1.3=py36h11da6c2_0
mistune=0.8.4=py36h0b31af3_1000
nbconvert=5.6.1=py36_0
nbformat=5.0.4=py_0
ncurses=6.1=h0a44026_1002
networkx=2.4=py_0
notebook=6.0.3=py36_0
numpy=1.18.1=py36hde6bac1_0
olefile=0.46=py_0
openssl=1.1.1d=h0b31af3_0
pandoc=2.9.2=0
pandocfilters=1.4.2=py_1
parso=0.6.1=py_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_1000
pillow=7.0.0=py36h918e99a_0
pip=20.0.2=py_2
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.3=py_0
protobuf=3.11.4=py36h4a8c4bd_0
ptyprocess=0.6.0=py_1001
pygments=2.5.2=py_0
pygpu=0.7.6=py36h3b54f70_1000
pyparsing=2.4.6=py_0
pyqt=5.6.0=py36hc26a216_1008
pyrsistent=0.15.7=py36h0b31af3_0
python=3.6.7=h4285619_1006
python-dateutil=2.8.1=py_0
pywavelets=1.1.1=py36h3b54f70_0
pyyaml=5.3=py36h0b31af3_0
pyzmq=19.0.0=py36h4bf09a9_0
qt=5.6.2=h822fa55_1013
qtconsole=4.7.1=py_0
qtpy=1.9.0=py_0
readline=8.0=hcfe32e1_0
scikit-image=0.16.2=py36h4f17bb1_0
scipy=1.3.1=py36hab3da7d_2
send2trash=1.5.0=py_0
setuptools=45.2.0=py36_0
sip=4.18.1=py36h0a44026_1000
six=1.14.0=py36_0
sqlite=3.30.1=h93121df_0
tapi=1000.10.8=ha1b3eb9_4
tensorboard=1.13.1=py36_0
tensorflow=1.13.2=hfddd6c2_0
tensorflow-base=1.13.2=py36_0
tensorflow-estimator=1.13.0=py36h24bf2e0_0
termcolor=1.1.0=py_2
terminado=0.8.3=py36_0
testpath=0.4.4=py_0
theano=1.0.4=py36h4a8c4bd_1001
tk=8.6.10=hbbe82c9_0
toolz=0.10.0=py_0
tornado=6.0.3=py36h0b31af3_4
traitlets=4.3.3=py36_0
wcwidth=0.1.8=py_0
webencodings=0.5.1=py_1
werkzeug=1.0.0=py_0
wheel=0.34.2=py_1
widgetsnbextension=3.5.1=py36_0
xz=5.2.4=h1de35cc_1001
yaml=0.2.2=h0b31af3_1
zeromq=4.3.2=h6de7cb9_2
zipp=3.0.0=py_0
zlib=1.2.11=h0b31af3_1006
zstd=1.4.4=he7fca8b_1
